==> 01 -- Assessing Data.ipynb <==
## Assessing Data
Analyzing fule efficiency data from 2008 and 2018, using data from Department of Energy: [Download Fuel Economy Data](https://www.fueleconomy.gov/feg/download.shtml/)

import pandas as pd
df_08 = pd.read_csv('all_alpha_08.csv')
df_18 = pd.read_csv('all_alpha_18.csv')

## How many rows (samples) and columns are in each data set?

# 2008
print(df_08.shape)
df_08.head(2)

# 2018
print(df_18.shape)
df_18.head(2)

## Number of duplicate rows in each dataset

# dupes in 2008
sum(df_08.duplicated())

# dupes in 2018
sum(df_18.duplicated())

## Datatypes of columns

# 2008
df_08.dtypes

# 2018
df_18.dtypes

> **The following features have different datatype in 2008 and 2018:**
> * Cyl (2018) - float
> * Cyl (2008) - string
> * Greenhouse Gas Score (2008) - string
> * Greenhouse Gas Score (2018) - int

## Features with missing values

# 2008
df_08.isnull().sum()

# 2018
df_18.isnull().sum()

## Number of non-null unique values for features in each dataset
Looking at Fuel, what are the unique values, and how many are there?

# 2008 count of non-null unique values
df_08.nunique()

# 2018 count of non-null unique values
df_18.nunique()

# 2008 unique values for Fuel
df_08.Fuel.unique()

# 2018 unique values for Fuel
df_18.Fuel.unique()
