==> 02 -- Cleaning Columns.ipynb <==
## Cleaning Columns and Column Labels
* Remove columns we won't use in our analysis
* Rename columns in two dataframes to match
* Replace spaces with _ in column names

import pandas as pd
df_08 = pd.read_csv('all_alpha_08.csv')
df_18 = pd.read_csv('all_alpha_18.csv')

### Check Data

# view 2008 dataset
df_08.head(1)

# view 2018 dataset
df_18.head(1)

> **Note:** 'Sales Area' in 2008 is 'Cert Region' in 2018

### Drop Unneeded Columns

# drop columns from 2008 dataset
df_08.drop(['Stnd', 'Underhood ID', 'FE Calc Appr', 'Unadj Cmb MPG'], axis=1, inplace=True)

# confirm changes
df_08.head(1)

# drop columns from 2018 dataset
df_18.drop(['Stnd', 'Stnd Description', 'Underhood ID', 'Comb CO2'], axis=1, inplace=True)


# confirm changes
df_18.head(1)

### Rename Columns for consistency across dataframes
Change 'Sales Area' feature in 2008 data to 'Cert Region' to match 2018 data

# rename Sales Area to Cert Region
df_08.rename(columns = {'Sales Area': 'Cert Region'}, inplace=True )

# confirm changes
df_08.head(1)

### Replace space with _ in column Names

# replace spaces with underscores and lowercase labels for 2008 dataset
df_08.rename(columns=lambda x: x.strip().lower().replace(" ", "_"), inplace=True)

# confirm changes
df_08.head(1)

# replace spaces with underscores and lowercase labels for 2018 dataset
df_18.rename(columns=lambda x: x.strip().lower().replace(" ", "_"), inplace=True)

# confirm changes
df_18.head(1)

# confirm column labels for 2008 and 2018 datasets are identical
df_08.columns == df_18.columns

# make sure they're all identical like this
(df_08.columns == df_18.columns).all()

# save new datasets for next section
df_08.to_csv('data_08_for_03.csv', index=False)
df_18.to_csv('data_18_for_03.csv', index=False)
