==> 03 -- filter, drop nulls, dedupe.ipynb <==
## Filter, drop nulls and dedupe
1. **Filter** For consistency, only compare cars certified by California standards. Filter both datasets using query to select only rows where cert_region is CA. Then, drop the cert_region columns, since it will no longer provide any useful information (we'll know every value is 'CA').
2. **Drop Nulls** Drop any rows in both datasets that contain missing values.
3. **Dedupe** Drop any duplicate rows in both datasets.

import pandas as pd
df_08 = pd.read_csv('data_08_for_03.csv')
df_18 = pd.read_csv('data_18_for_03.csv')

# view dimensions of dataset 2008
df_08.shape

# view dimensions of dataset 2018
df_18.shape

### Filter By Certification Region
Just want to look at cars that meet California certifcation standards

# filter datasets for rows following California standards
df_08 = df_08.query('cert_region == "CA"')
df_18 = df_18.query('cert_region == "CA"')

# confirm only certification region is California in 2008 data
df_08['cert_region'].unique()

# confirm only certification region is California in 2018 data
df_18['cert_region'].unique()

# drop certification region columns form both datasets
df_08.drop(['cert_region'], axis=1, inplace=True)
df_18.drop(['cert_region'], axis=1, inplace=True)

df_08.shape

df_18.shape

### Drop Rows with Missing Values

# view missing value count for each feature in 2008
df_08.isnull().sum()

# view missing value count for each feature in 2018
df_18.isnull().sum()

# drop rows with any null values in both datasets
df_08.dropna(inplace=True)
df_18.dropna(inplace=True)

# checks if any of columns in 2008 have null values - should print False
df_08.isnull().sum().any()

# checks if any of columns in 2018 have null values - should print False
df_18.isnull().sum().any()

### Dedupe Data

# print number of duplicates in 2008 and 2018 datasets
print(sum(df_08.duplicated()))
print(sum(df_18.duplicated()))

# drop duplicates in both datasets
df_08.drop_duplicates(inplace = True)
df_18.drop_duplicates(inplace = True)

# print number of duplicates again to confirm dedupe - should both be 0
print(sum(df_08.duplicated()))
print(sum(df_18.duplicated()))

# save new datasets for next section
df_08.to_csv('data_08_for_04.csv', index=False)
df_18.to_csv('data_18_for_04.csv', index=False)

df_08.shape

df_18.shape
