==> 04 -- Inspecting Data Types.ipynb <==
## Inspecting Data Types

import pandas as pd
df_08 = pd.read_csv('data_08_for_04.csv')
df_18 = pd.read_csv('data_18_for_04.csv')

### Let's look at the `cyl` feature

# in 2008
df_08['cyl'].sample(5)

# in 2018
df_18['cyl'].sample(5)

> `cyl` is string in 2018 and Float in 2018

### Let's look at the `air_pollution_score` feature

# check the type of one sample of air_pollution_score
type(df_08['air_pollution_score'][0])

# check the type of one sample of air_pollution_score
type(df_18['air_pollution_score'][0])

# is the data the same type across the two dataframes?
type(df_08['air_pollution_score'][0]) == type(df_18['air_pollution_score'][0])

## How about for some other columns?

# first, what columns are there?
df_08.columns

### MPG data
* city_mpg
* hwy_mpg
* cmb_mpg

# Let's look at some data for 2008
df_08.city_mpg.sample(5)

# Let's look at some data for 2018
df_18.city_mpg.sample(5)

# what is type of data for city_mpg column?
type_08 = type(df_08.city_mpg[0])
print('2008 data, city_mpg column: {}'.format(type_08))
type_18 = type(df_18.city_mpg[0])
print('2018 data, city_mpg column: {}'.format(type_18))
type_08 == type_18

# what is type of data for city_mpg column?
type_08 = type(df_08.hwy_mpg[0])
print('2008 data, hwy_mpg column: {}'.format(type_08))
type_18 = type(df_18.city_mpg[0])
print('2018 data, hwy_mpg column: {}'.format(type_18))
type_08 == type_18

# what is type of data for city_mpg column?
type_08 = type(df_08.cmb_mpg[0])
print('2008 data, cmb_mpg column: {}'.format(type_08))
type_18 = type(df_18.cmb_mpg[0])
print('2018 data, cmb_mpg column: {}'.format(type_18))
type_08 == type_18

### Check out `greenhouse_gas_score`

# let's look at some data for 2008
df_08.greenhouse_gas_score.sample(5)

# let's look at some data for 2018
df_18.greenhouse_gas_score.sample(5)

# what data type in 2008?
type(df_08.greenhouse_gas_score[0])

# what data type in 2018?
type(df_18.greenhouse_gas_score[0])

# are they the same?
type(df_08.greenhouse_gas_score[0]) == type(df_18.greenhouse_gas_score[0])
