==> 05 -- Fix datatypes.ipynb <==
import pandas as pd
df_08 = pd.read_csv('data_08_for_04.csv') # can use same files as last exercise. No data was changed
df_18 = pd.read_csv('data_18_for_04.csv')

## Fix cyl datatype
* 2008: extract int from string.
* 2018: convert float to int.

# check value counts for the 2008 cyl column
df_08['cyl'].value_counts()

### Extract Numbers from Strings in 2008 Data
Using regex extract one or more numbers as cast as int

df_08.cyl = df_08.cyl.str.extract('(\d+)').astype(int)

# check type
type(df_08.cyl[0])

# check values
df_08['cyl'].value_counts()

# convert 2018 cyl data to int
df_18['cyl'] = df_18['cyl'].astype(int)

# check type
type(df_18.cyl[0])

# are 2008 and 2018 the same type in cyl?
type(df_08.cyl[0]) == type(df_18.cyl[0])

## Fix air_pollution_score datatype
* 2008: convert string to float.
* 2018: convert int to float.

# check value counts for the 2008 cyl column
df_08['air_pollution_score'].value_counts()

# check value counts for the 2018 air_pollution_score column
df_18['air_pollution_score'].value_counts()

**Method we used above won't work for 2008 data**

# convert to float
df_08['air_pollution_score'] = df_08['air_pollution_score'].astype(float)

# Figuring out the issue
Looks like this isn't going to be as simple as converting the datatype.
There's a string value `6/4` in the data somwhere, but where?

# To find out the position for the value - "6/4"
df_08.query('air_pollution_score == "6/4"')

# at index 582
df_08['air_pollution_score'][582]

## Hybrid car values are stored as strings with fuel estimates for each fuel type
According to [this link](http://www.fueleconomy.gov/feg/findacarhelp.shtml#airPollutionScore)from the PDF documentation:

    "If a vehicle can operate on more than one type of fuel, an estimate is provided for each fuel type."

The mpg columns and greenhouse gas scores also seem to have the same problem. All vehicles with more than one fuel type, or hybrids, like the one above (it uses ethanol AND gas) will have a string that holds two values - one for each fuel.

### Solution
Create two entries for each hybrid vehicle: one row with one fuel and the values for that fuel type, and a second row with values for the second fuel type.

# First, let's get all the hybrids in 2008
hb_08 = df_08[df_08['fuel'].str.contains('/')]
hb_08
# there's only 1

# hybrids in 2018
hb_18 = df_18[df_18['fuel'].str.contains('/')]
hb_18

# create two copies of the 2008 hybrids dataframe
df1 = hb_08.copy()  # data on first fuel type of each hybrid vehicle
df2 = hb_08.copy()  # data on second fuel type of each hybrid vehicle

# Each one should look like this
df1

df2

# columns to split by "/"
split_columns = ['fuel', 'air_pollution_score', 'city_mpg', 'hwy_mpg', 'cmb_mpg', 'greenhouse_gas_score']

# apply split function to each column of each dataframe copy
for c in split_columns:
    df1[c] = df1[c].apply(lambda x: x.split("/")[0])
    df2[c] = df2[c].apply(lambda x: x.split("/")[1])

# this dataframe holds info for the FIRST fuel type of the hybrid
# aka the values before the "/"s (ethanol in this case)
df1

# this dataframe holds info for the SECOND fuel type of the hybrid
# aka the values before the "/"s (gas in this case)
df2

# combine dataframes to add to the original dataframe
new_rows = df1.append(df2)

# now we have separate rows for each fuel type of each vehicle!
new_rows

# drop the original hybrid rows
df_08.drop(hb_08.index, inplace=True)

# add in our newly separated rows
df_08 = df_08.append(new_rows, ignore_index=True)

# check that all the original hybrid rows with "/"s are gone
df_08[df_08['fuel'].str.contains('/')]

df_08.shape

### Fix 2018 Data
# create two copies of the 2088 hybrids dataframe
df1 = hb_18.copy()  # data on first fuel type of each hybrid vehicle
df2 = hb_18.copy()  # data on second fuel type of each hybrid vehicle

# Each dataframe should look like this
df1.head(5)

# columns to split by "/"
# not required for air_pollution_score or greenhouse_gas_score here 
# because these columns are already ints in the 2018 dataset.
split_columns = ['fuel', 'city_mpg', 'hwy_mpg', 'cmb_mpg']

# apply split function to each column of each dataframe copy
for c in split_columns:
    df1[c] = df1[c].apply(lambda x: x.split("/")[0])
    df2[c] = df2[c].apply(lambda x: x.split("/")[1])

# combine dataframes to add to the original dataframe
new_rows = df1.append(df2)

# now we have separate rows for each fuel type of each vehicle!
new_rows.head(5)

# drop the original hybrid rows
df_18.drop(hb_18.index, inplace=True)

# add in our newly separated rows
df_18 = df_18.append(new_rows, ignore_index=True)

# check that all the original hybrid rows with "/"s are gone
df_18[df_18['fuel'].str.contains('/')]

df_18.shape

### Convert to air_pollution_score to Floats

df_08.air_pollution_score = df_08.air_pollution_score.astype(float)
type(df_08.air_pollution_score[0])

df_18.air_pollution_score = df_18.air_pollution_score.astype(float)
type(df_18.air_pollution_score[0])

### Convert `city_mpg`, `hwy_mpg`, `cmb_mpg` datatypes to floats

mpg_columns = ['city_mpg', 'hwy_mpg', 'cmb_mpg']
for c in mpg_columns:
    df_08[c] = df_08[c].astype(float)
    df_18[c] = df_18[c].astype(float)
    print('2008 {} type: {}'.format(c,type(df_08[c][0])))
    print('2018 {} type: {}'.format(c,type(df_18[c][0])))

## Fix `greenhouse_gas_score` datatype
* 2008: convert from float to int

df_08.greenhouse_gas_score = df_08.greenhouse_gas_score.astype(int)

### Final Confirmation

df_08.dtypes

df_18.dtypes

df_08.dtypes == df_18.dtypes

# Save your new CLEAN datasets as new files!
df_08.to_csv('clean_08.csv', index=False)
df_18.to_csv('clean_18.csv', index=False)
